Move emoji and noise go files from their own separate packages into one common package for better readability. Refeactor main.go to reflect this
